+++ formTypes/method.js
@@ -29,11 +29,8 @@ AutoForm.addFormType('method', {
         // Call the method. If a ddp connection was provided, use
         // that instead of the default Meteor connection
         var ddp = c.formAttributes.ddp;
-        if (ddp && ddp.call && typeof ddp.call === 'function') {
-          ddp.call(c.formAttributes.meteormethod, doc, c.result);
-        } else {
-          Meteor.call(c.formAttributes.meteormethod, doc, c.result);
-        }
+        ddp = (ddp && typeof ddp.call === 'function') ? ddp : Meteor;
+        ddp.call(c.formAttributes.meteormethod, doc, c.result);
       }
     });
   },