+++ inputTypes/boolean-checkbox/boolean-checkbox.js
@@ -4,38 +4,10 @@ AutoForm.addInputType("boolean-checkbox", {
     return !!this.is(":checked");
   },
   valueConverters: {
-    "string": function (val) {
-      if (val === true) {
-        return "TRUE";
-      } else if (val === false) {
-        return "FALSE";
-      }
-      return val;
-    },
-    "stringArray": function (val) {
-      if (val === true) {
-        return ["TRUE"];
-      } else if (val === false) {
-        return ["FALSE"];
-      }
-      return val;
-    },
-    "number": function (val) {
-      if (val === true) {
-        return 1;
-      } else if (val === false) {
-        return 0;
-      }
-      return val;
-    },
-    "numberArray": function (val) {
-      if (val === true) {
-        return [1];
-      } else if (val === false) {
-        return [0];
-      }
-      return val;
-    }
+    "string": AutoForm.valueConverters.booleanToString,
+    "stringArray": AutoForm.valueConverters.booleanToStringArray,
+    "number": AutoForm.valueConverters.booleanToNumber,
+    "numberArray": AutoForm.valueConverters.booleanToNumberArray
   },
   contextAdjust: function (context) {
     if (context.value === true) {