+++ inputTypes/number/number.js
@@ -1,43 +1,14 @@
 AutoForm.addInputType("number", {
   template: "afInputNumber",
   valueOut: function () {
-    return AutoForm.Utility.stringToNumber(this.val());
+    return AutoForm.valueConverters.stringToNumber(this.val());
   },
   valueConverters: {
-    "string": function (val) {
-      if (typeof val === "number") {
-        return val.toString();
-      }
-      return val;
-    },
-    "stringArray": function (val) {
-      if (typeof val === "number") {
-        return [val.toString()];
-      }
-      return val;
-    },
-    "numberArray": function (val) {
-      if (typeof val === "number") {
-        return [val];
-      }
-      return val;
-    },
-    "boolean": function (val) {
-      if (val === 0) {
-        return false;
-      } else if (val === 1) {
-        return true;
-      }
-      return val;
-    },
-    "booleanArray": function (val) {
-      if (val === 0) {
-        return [false];
-      } else if (val === 1) {
-        return [true];
-      }
-      return val;
-    }
+    "string": AutoForm.valueConverters.numberToString,
+    "stringArray": AutoForm.valueConverters.numberToStringArray,
+    "numberArray": AutoForm.valueConverters.numberToNumberArray,
+    "boolean": AutoForm.valueConverters.numberToBoolean,
+    "booleanArray": AutoForm.valueConverters.numberToBooleanArray
   },
   contextAdjust: function (context) {
     if (typeof context.atts.max === "undefined" && typeof context.max === "number") {