+++ inputTypes/radio/radio.js
@@ -6,12 +6,7 @@ AutoForm.addInputType("radio", {
     }
   },
   valueConverters: {
-    "stringArray": function (val) {
-      if (typeof val === "string" && val.length > 0) {
-        return [val];
-      }
-      return val;
-    }
+    "stringArray": AutoForm.valueConverters.stringToStringArray
   }
 });