+++ inputTypes/select-checkbox-inline/select-checkbox-inline.js
@@ -35,7 +35,7 @@ AutoForm.addInputType("select-checkbox-inline", {
   }
 });
 
-Template["afCheckboxGroupInline"].helpers({
+Template.afCheckboxGroupInline.helpers({
   atts: function selectedAttsAdjust() {
     var atts = _.clone(this.atts);
     if (this.selected) {
@@ -49,6 +49,6 @@ Template["afCheckboxGroupInline"].helpers({
   dsk: function dsk() {
     return {
       "data-schema-key": this.atts["data-schema-key"]
-    }
+    };
   }
 });