+++ inputTypes/week/week.js
@@ -1,11 +1,6 @@
 AutoForm.addInputType("week", {
   template: "afInputWeek",
   valueConverters: {
-    "stringArray": function (val) {
-      if (typeof val === "string" && val.length > 0) {
-        return [val];
-      }
-      return val;
-    }
+    "stringArray": AutoForm.valueConverters.stringToStringArray
   }
 });