+++ templates/bootstrap3-horizontal/components/quickForm/quickForm.js
@@ -5,10 +5,26 @@ Template['quickForm_bootstrap3-horizontal'].helpers({
   labelClass: function () {
     return this.atts["label-class"];
   },
+  fieldGroupLabel: function () {
+    var name = this.name;
+
+    // if field group name is of the form XY_abcde where "XY" is a number, remove prefix
+    if (!isNaN(parseInt(name.substr(0,2), 10)) && name.charAt(2) === "_") {
+      name = name.substr(3);
+    }
+
+    // if SimpleSchema.defaultLabel is defined, use it
+    if (typeof SimpleSchema.defaultLabel === "function") {
+      return SimpleSchema.defaultLabel(name);
+    } else {
+      // else, just capitalise name
+      return name.charAt(0).toUpperCase() + name.slice(1);
+    }
+  },
   quickFieldsAtts: function () {
     // These are the quickForm attributes that we want to forward to
     // the afQuickFields component.
-    return _.pick(this.atts, 'id-prefix', 'input-col-class', 'label-class');
+    return _.pick(this.atts, 'fields', 'id-prefix', 'input-col-class', 'label-class');
   },
   submitButtonAtts: function () {
     var qfAtts = this.atts;
@@ -22,10 +38,15 @@ Template['quickForm_bootstrap3-horizontal'].helpers({
   },
   qfAutoFormContext: function () {
     var ctx = _.clone(this.qfAutoFormContext || {});
-    ctx = AutoForm.Utility.addClass(ctx, "form-horizontal");
-    delete ctx["input-col-class"];
-    delete ctx["label-class"];
-    delete ctx["id-prefix"];
+    ctx = AutoForm.Utility.addClass(ctx, 'form-horizontal');
+
+    // input-col-class and label-class attributes are unique to this template so they will
+    // not have been removed by AutoForm core. We remove them from the autoForm context
+    // because they are attributes supported only by quickFields, quickField,
+    // afObjectField, afArrayField, and afFormGroup.
+    delete ctx['input-col-class'];
+    delete ctx['label-class'];
+
     return ctx;
   }
 });